Copy IDC for goldfish_rotary to /system/usr/idc
This IDC file won't be used by the emulator, unless there's an input
device called goldfish_rotary. That device is being added to the kernel
in http://r.android.com/321420.

Once that device is present, EventHub will look for a file named
goldfish_rotary.idc in the system/usr/idc directory to get the device
configuration.

Bug: 33292948
Change-Id: I384f29980d802280e626915844bddc49bc46110d